isg-android-control/android-control/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyYAML==6.0.2
redis==5.0.4
paho-mqtt==1.6.1
Pillow==10.4.0
